requirements.txt
matplotlib==3.1.1
numpy==1.17.2
opencv-contrib-python-headless==4.1.1.26
Pillow==6.1.0
PyYAML==5.1.2
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.1
tensorboard==2.0.0
torch==1.4.0
torchvision==0.5.0